+++ api/src/Article/Service/ArticleManager.php
@@ -28,6 +28,7 @@ use Doctrine\ORM\EntityManagerInterface;
 use Psr\Log\LoggerInterface;
 use App\Article\Entity\Paragraph;
 use App\Article\Entity\Article as ArticleEntity;
+use App\Article\Entity\Iframe;
 use App\Article\Entity\RssElement;
 use App\Article\Repository\SectionRepository;
 use App\Article\Repository\ParagraphRepository;
@@ -69,14 +70,25 @@ class ArticleManager
 
     private $articleFeed;
     private $articleFeedNumber;
+    private $articleIframeMaxWidth;
+    private $articleIframeMaxHeight;
 
     /**
      * Constructor.
      *
      * @param EntityManagerInterface $entityManager
      */
-    public function __construct(EntityManagerInterface $entityManager, LoggerInterface $logger, SectionRepository $sectionRepository, ParagraphRepository $paragraphRepository, ArticleRepository $articleRepository, $articleFeed, $articleFeedNumber)
-    {
+    public function __construct(
+        EntityManagerInterface $entityManager,
+        LoggerInterface $logger,
+        SectionRepository $sectionRepository,
+        ParagraphRepository $paragraphRepository,
+        ArticleRepository $articleRepository,
+        string $articleFeed,
+        int $articleFeedNumber,
+        int $articleIframeMaxWidth,
+        int $articleIframeMaxHeight
+    ) {
         $this->entityManager = $entityManager;
         $this->logger = $logger;
         $this->sectionRepository = $sectionRepository;
@@ -84,6 +96,8 @@ class ArticleManager
         $this->articleRepository = $articleRepository;
         $this->articleFeed = $articleFeed;
         $this->articleFeedNumber = $articleFeedNumber;
+        $this->articleIframeMaxWidth = $articleIframeMaxWidth;
+        $this->articleIframeMaxHeight = $articleIframeMaxHeight;
     }
 
     /**
@@ -190,6 +204,11 @@ class ArticleManager
         $article->setDescription($rssElement->getDescription());
         $article->setImage($rssElement->getImage());
         $article->setPubDate($rssElement->getPubDate());
+        
+        if (!is_null($rssElement->getIframe())) {
+            $rssIframe = $rssElement->getIframe();
+            $article->setIFrame('<iframe src="'.$rssIframe->getSrc().'" allowfullscreen title="'.$rssIframe->getTitle().'" width="'.$rssIframe->getWidth().'" height="'.$rssIframe->getHeight().'" allow="'.$rssIframe->getAllow().'" frameborder="0"></iframe>');
+        }
 
         return $article;
     }
@@ -247,6 +266,30 @@ class ArticleManager
                 $image = $dom->getElementsByTagName('img')->item(0)->getAttribute('src');
                 $rssElement->setImage($image);
             }
+
+            // If there is an iframe (usually a video) we parse it. We only take the first one.
+            if ($dom->getElementsByTagName('iframe')->length > 0) {
+                $iframe = new Iframe();
+                $iframe->setSrc($dom->getElementsByTagName('iframe')->item(0)->getAttribute('src'));
+                $iframe->setTitle($dom->getElementsByTagName('iframe')->item(0)->getAttribute('title'));
+                
+                if ((int)$dom->getElementsByTagName('iframe')->item(0)->getAttribute('width') > $this->articleIframeMaxWidth) {
+                    $iframe->setWidth($this->articleIframeMaxWidth);
+                } else {
+                    $iframe->setWidth($dom->getElementsByTagName('iframe')->item(0)->getAttribute('width'));
+                }
+                
+                if ((int)$dom->getElementsByTagName('iframe')->item(0)->getAttribute('height') > $this->articleIframeMaxHeight) {
+                    $iframe->setHeight($this->articleIframeMaxHeight);
+                } else {
+                    $iframe->setHeight($dom->getElementsByTagName('iframe')->item(0)->getAttribute('height'));
+                }
+
+                $iframe->setAllow($dom->getElementsByTagName('iframe')->item(0)->getAttribute('allow'));
+                
+                $rssElement->setIframe($iframe);
+            }
+
             $counter++;
             if ($counter == $articleFeedNumber) {
                 break;
@@ -281,10 +324,10 @@ class ArticleManager
             $rssItems = $this->getRssFeeds();
 
             foreach ($rssItems as $rssItem) {
-                $article[] = $this->makeArticleFromRss($rssItem);
+                $articles[] = $this->makeArticleFromRss($rssItem);
             }
         }
 
-        return $article;
+        return $articles;
     }
 }